using uint16, and rely on same ordering

+++ go/libraries/doltcore/schema/schema.go
@@ -71,10 +71,10 @@ type Schema interface {
 
 	// GetPkPrefixLengths returns a slice of pairs (tag, prefix length) containing the prefix lengths
 	// for any columns that use a text/blob index
-	GetPkPrefixLengths() []uint64
+	GetPkPrefixLengths() []uint16
 
 	// SetPkPrefixLengths specifies the tags and prefix lengths for the columns that use text/blob indexes
-	SetPkPrefixLengths([]uint64) error
+	SetPkPrefixLengths([]uint16) error
 
 	// AddColumn adds a column to this schema in the order given and returns the resulting Schema.
 	// The new column cannot be a primary key. To alter primary keys, create a new schema with those keys.
